requirements.txt
Django>=2.1.3,<2.2.0
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
